[Fix] UI Fix (#48)
