+++ for.JOY/for.JOY/Extensions/View+.swift
@@ -8,7 +8,22 @@
 import SwiftUI
 
 extension View {
-    func keyboardAdaptive() -> some View {
-        ModifiedContent(content: self, modifier: KeyboardAdaptiveModifier())
+    func navigationBarColor(_ backgroundColor: UIColor?) -> some View {
+       self.modifier(NavigationBarColorModifier(backgroundColor: backgroundColor))
+   }
+    
+    func snapshot() -> UIImage {
+        let controller = UIHostingController(rootView: self)
+        let view = controller.view
+
+        let targetSize = CGSize(width: 346, height: 459)
+        view?.bounds = CGRect(origin: .zero, size: targetSize)
+        view?.backgroundColor = .clear
+
+        let renderer = UIGraphicsImageRenderer(size: targetSize)
+
+        return renderer.image { _ in
+            view?.drawHierarchy(in: controller.view.bounds, afterScreenUpdates: true)
+        }
     }
 }
