+++ for.JOY/for.JOY/for.JOY.entitlements
@@ -4,7 +4,7 @@
 <dict>
 	<key>com.apple.security.application-groups</key>
 	<array>
-		<string>group.newyork.forJoy</string>
+		<string>group.for.helia</string>
 	</array>
 </dict>
 </plist>
